Add more classes
