+++ src/main/java/TaskList.java
@@ -1,42 +1,32 @@
 import java.util.ArrayList;
-import java.util.Scanner;
 
 public class TaskList {
     ArrayList<Task> fullList;
+    Ui ui;
     private final static String line = "------------------------------------";
 
     public TaskList() {
         this.fullList = new ArrayList<>();
+        this.ui = new Ui();
     }
 
     public TaskList(ArrayList<Task> fullList) {
         this.fullList = fullList;
+        this.ui = new Ui();
     }
 
     public void addToList(Task task) {
         this.fullList.add(task);
-        System.out.println(line);
-        System.out.println("Got it. I've added this task:");
-        System.out.println(task.toString());
-        System.out.println("Now you have " + fullList.size() + " tasks in the list.");
-        System.out.println(line);
+        ui.showAddMessage(task, this.fullList.size());
     };
 
     public void deleteFromList(int i) {
         if (i >= 0 && i < fullList.size()) {
             Task item = fullList.get(i);
             this.fullList.remove(item);
-
-            System.out.println(line + "\n" +
-                    "Noted. I've removed this task:\n" +
-                    item.toString() + "\n" +
-                    "Now you have " + this.fullList.size() + " tasks in the list."
-                    + "\n" + line
-            );
+            ui.showDeleteMessage(item, this.fullList.size());
         } else {
-            System.out.println(line);
-            System.out.println("No such item exists");
-            System.out.println(line);
+            ui.showNoItemMessage();
         }
     }
 
@@ -44,15 +34,9 @@ public class TaskList {
         if (index >= 0 && index < fullList.size()) {
             Task curr = fullList.get(index);
             curr.markDone();
-
-            System.out.println(line);
-            System.out.println("Nice! I've marked this task as done:");
-            System.out.println(curr.toString());
-            System.out.println(line);
+            ui.showMarkMessage(curr);
         } else {
-            System.out.println(line);
-            System.out.println("No such item exists");
-            System.out.println(line);
+            ui.showNoItemMessage();
         }
     }
 
@@ -60,15 +44,9 @@ public class TaskList {
         if (index >= 0 && index < fullList.size()) {
             Task curr = fullList.get(index);
             curr.markNotDone();
-
-            System.out.println(line);
-            System.out.println("Ok, I've marked this task as not done yet:");
-            System.out.println(curr.toString());
-            System.out.println(line);
+            ui.showUnmarkMessage(curr);
         } else {
-            System.out.println(line);
-            System.out.println("No such item exists");
-            System.out.println(line);
+            ui.showNoItemMessage();
         }
     }
